refactor(scan): decompose RegexCodeSlicer (#302)
refactor(scan): decompose RegexCodeSlicer into focused collaborators

RegexCodeSlicer had grown to 465 lines mixing line classification,
block-comment stripping, paren-continuation tracking, heredoc
handling, and orchestration in one class - a SOLID/SRP violation
that made each concern harder to test and reason about in isolation.

Split into:
- StringLiteralMasker: neutralizes string-literal content
- SecurityRelevantLineClassifier: structural/security-token retention
- BlockCommentStripper: strips block comments, tracks open/close state
- ParenContinuationTracker: tracks open-paren depth across lines
- HeredocLineTracker: retains a heredoc/nowdoc body verbatim

Retention itself is now a LineRetentionDeciderInterface.
ContinuationForcedLineRetentionDecider decorates
SecurityRelevantLineClassifier so a line continuing an open construct
is retained regardless of the base classifier's verdict, replacing
the inline three-way boolean OR that used to live in slice().

RegexCodeSlicer is now a 93-line orchestrator delegating to these
collaborators. Purely internal (@internal, not part of the BC
promise) with identical output for every existing case - the
existing 85-test suite passes unchanged, and diff-scoped Infection
against fix/regex-code-slicer-recursion-segfault kills all 118
mutants (100% MSI) across the 9 changed/added files.
