+++ src/Audit/Infrastructure/Scan/LineRetentionContext.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Scan;
+
+/**
+ * @internal not part of the BC promise — see docs/versioning.md
+ */
+final readonly class LineRetentionContext
+{
+    public function __construct(
+        public int $openParenDepth,
+        public bool $opensHeredoc,
+    ) {}
+}
