+++ src/Audit/Infrastructure/Scan/LineRetentionDeciderInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Scan;
+
+/**
+ * @internal not part of the BC promise — see docs/versioning.md
+ *
+ * Decides whether a single line of a sliced file is kept verbatim or elided.
+ */
+interface LineRetentionDeciderInterface
+{
+    public function isRetained(string $line, LineRetentionContext $lineRetentionContext): bool;
+}
